hotfix for potential issue with conflicts on existing workspaces accidentally deleting them

+++ app/models/study.rb
@@ -1516,8 +1516,15 @@ class Study
         rescue => e
           # delete workspace on any fail as this amounts to a validation fail
           Rails.logger.info "#{Time.now}: Error creating workspace: #{e.message}"
-          Study.firecloud_client.delete_workspace(self.firecloud_workspace)
-          errors.add(:firecloud_workspace, " creation failed: #{e.message}; Please try again later.")
+          # delete firecloud workspace unless error is 409 Conflict (workspace already taken)
+          if e.message != '409 Conflict'
+            Study.firecloud_client.delete_workspace(self.firecloud_workspace)
+            errors.add(:firecloud_workspace, " creation failed: #{e.message}; Please try again.")
+          else
+            errors.add(:firecloud_workspace, ' - there is already an existing workspace using this name.  Please choose another name for your study.')
+            errors.add(:name, ' - you must choose a different name for your study.')
+            self.firecloud_workspace = nil
+          end
           return false
         end
       end
